Added global exception handling for Book APIs

+++ src/main/java/com/ketan/library/service/BookService.java
@@ -1,11 +1,13 @@
 package com.ketan.library.service;
 
 import com.ketan.library.entity.Book;
+import com.ketan.library.exception.BookNotFoundException;
 import com.ketan.library.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookService {
@@ -21,33 +23,45 @@ public class BookService {
     }
 
     public Book getBookById(int id) {
-        return bookRepository.findById(id).orElse(null);
-    }
+        Optional<Book> optionalBook = bookRepository.findById(id);
 
+        if (optionalBook.isPresent()) {
+            return optionalBook.get();
+        } else {
+            throw new BookNotFoundException("Book not found with id: " + id);
+        }
+    }
     public String deleteBook(int id) {
-        bookRepository.deleteById(id);
+      Optional<Book>optionalBook = bookRepository.findById(id);
+      if(optionalBook.isPresent()){
+          bookRepository.deleteById(id);
+      }else{
+          throw new BookNotFoundException("Book not found with id: "+id);
+      }
         return "Book deleted successfully";
     }
 
     public Book updateBook(int id, Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
+        Optional<Book> optionalBook = bookRepository.findById(id);
 
-        if (existingBook != null) {
+        if (optionalBook.isPresent()) {
+            Book existingBook = optionalBook.get();
+
             existingBook.setTitle(book.getTitle());
             existingBook.setAuthor(book.getAuthor());
             existingBook.setPrice(book.getPrice());
 
             return bookRepository.save(existingBook);
+        } else {
+            throw new BookNotFoundException("Book not found with id: " + id);
         }
-
-        return null;
     }
 
-
     public Book updateBookPartial(int id, Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
+        Optional<Book> optionalBook = bookRepository.findById(id);
 
-        if (existingBook != null) {
+        if (optionalBook.isPresent()) {
+            Book existingBook = optionalBook.get();
 
             if (book.getTitle() != null) {
                 existingBook.setTitle(book.getTitle());
@@ -62,8 +76,8 @@ public class BookService {
             }
 
             return bookRepository.save(existingBook);
+        } else {
+            throw new BookNotFoundException("Book not found with id: " + id);
         }
-
-        return null;
     }
 }
